# Optimize explosion entity raycasts