--- a/common/src/main/java/net/caffeinemc/mods/lithium/mixin/world/explosions/entity_raycast/package-info.java
+++ b/common/src/main/java/net/caffeinemc/mods/lithium/mixin/world/explosions/entity_raycast/package-info.java
@@ -1,0 +1,4 @@
+@MixinConfigOption(description = "Various improvements to explosion entity damage, e.g. simplifying the raycasts.")
+package net.caffeinemc.mods.lithium.mixin.world.explosions.entity_raycast;
+
+import net.caffeinemc.gradle.MixinConfigOption;