## Overview

This pull request does two things:
1. Fixes the [unused `HashSet` allocation redirect](https://github.com/CaffeineMC/lithium/blob/947217c2908f81c97e864a85f1dface7b00fac2e/common/src/main/java/net/caffeinemc/mods/lithium/mixin/world/explosions/block_raycast/ServerExplosionMixin.java#L106-L112) in `block_raycast.ServerExplosionMixin` that was removed during the 1.21.3 update of lithium (`NEW` redirects *have* to return a non-null object.)
2. Readds and improves upon the specialized raycast hit factory that I had added in my earlier pull request (#507)

## Benchmarks
~80% improvement in time spent exploding 4096 tnts.
Benchmarked using seed `-3199243952349285632` (randomly chosen) and 4096 tnt (`/fill ~1 ~-7 ~ ~16 ~8 ~15 tnt`)

Before:
6612 ms spent in `World#createExplosion` (5102 ms spent in `ExplosionImpl#damageEntities`)
![image](https://github.com/user-attachments/assets/97d8d6b0-bfdc-4da5-82be-5383068d2c14)

After:
3710 ms spent in `World#createExplosion` (1982 ms spent in `ExplosionImpl#damageEntities`)
![image](https://github.com/user-attachments/assets/f807f8e9-a39c-4680-8473-038d0887bb03)