@2No2Name 2025-02-23
The `BlockGetter.clipWithInteractionOverride` is replaced with `VoxelShape.clip` because the raycast result's direction on the block does not matter for explosions