@2No2Name 2025-03-18
This PR causes https://github.com/CaffeineMC/lithium/issues/630 due to removing the EntityCollisionContext, which affects raycasts, e.g. through scaffolding blocks. This is fixed in 0.15.1